requirements.txt
streamlit
requests
pdf2image
langchain
langchain-community
langchain-huggingface
langchain-ollama
streamlit-mic-recorder
pyttsx3
faiss-cpu
sentence-transformers
python-docx
nltk

# Optional: Anthropic (for Claude / Claude Haiku models)
anthropic
